backend-render/requirements.txt
numpy==1.21.0
tensorflow==2.5.0
Pillow==8.3.1
scikit-learn==0.24.2
gunicorn==20.1.0 